Temporal.Core.Client

Connecting to the server

data Client

A client connection to the Temporal server.

The client is thread-safe and can be shared across threads.

Clients are expensive to create, so you should generally create one per process and share it across your application..

The client doesn't have an explicit close method, but will be cleaned up when it is garbage collected.

clientConfig :: Client -> ClientConfig

connectClient :: Runtime -> ClientConfig -> IO Client

Connect to the Temporal server using a given runtime.

Throws ClientConnectionError if the connection fails.

defaultClientConfig :: ClientConfig

closeClient :: Client -> IO ()

clientRuntime :: Client -> Runtime

touchClient :: Client -> IO ()

data CoreClient

data ClientConfig

Constructors

ClientConfig

Fields

targetUrl :: Text

clientName :: Text

clientVersion :: Text

metadata :: HashMap Text Text

identity :: Text

tlsConfig :: Maybe ClientTlsConfig

retryConfig :: Maybe ClientRetryConfig

data ClientTlsConfig

Constructors

ClientTlsConfig

Fields

serverRootCaCert :: Maybe ByteVector

domain :: Maybe Text

clientCert :: Maybe ByteVector

clientPrivateKey :: Maybe ByteVector

newtype ByteVector

Constructors

ByteVector

Fields

byteVector :: ByteString

data ClientRetryConfig

Constructors

ClientRetryConfig

Fields

initialIntervalMillis :: Word64

randomizationFactor :: Double

multiplier :: Double

maxIntervalMillis :: Word64

maxElapsedTimeMillis :: Maybe Word64

maxRetries :: Word64

Making calls to the server

call :: forall svc (t :: Symbol). HasMethodImpl svc t => PrimRpcCall -> Client -> MethodInput svc t -> IO (Either RpcError (MethodOutput svc t))

data RpcCall a

Constructors

RpcCall

Fields

req :: a

retry :: Bool

metadata :: HashMap Text Text

timeoutMillis :: Maybe Word64

data RpcError

Constructors

RpcError

Fields

code :: Word32

message :: Text

details :: ByteString

data ClientError

Constructors

ClientConnectionError Text

ClientClosedError

Primitive access

data CRpcCall

type TokioCall e a = StablePtr PrimMVar -> Int -> TokioSlot e -> TokioSlot a -> IO ()

type TokioResult a = ForeignPtr (Ptr a)

type PrimRpcCall = Ptr CoreClient -> Ptr CRpcCall -> TokioCall CRPCError (CArray Word8)

Use the underlying Rust client pointer

withClient :: Client -> (Ptr CoreClient -> IO a) -> IO a