Temporal.Core.Worker

Documentation

data Worker (ty :: WorkerType)

getWorkerClient :: Worker 'Real -> Client

data WorkerConfig

Constructors

WorkerConfig

Fields

namespace :: Text

taskQueue :: Text

buildId :: Text

identityOverride :: Maybe Text

maxCachedWorkflows :: Word64

maxOutstandingWorkflowTasks :: Word64

maxOutstandingActivities :: Word64

maxOutstandingLocalActivities :: Word64

maxConcurrentWorkflowTaskPolls :: Word64

nonstickyToStickyPollRatio :: Float

maxConcurrentActivityTaskPolls :: Word64

noRemoteActivities :: Bool

stickyQueueScheduleToStartTimeoutMillis :: Word64

maxHeartbeatThrottleIntervalMillis :: Word64

defaultHeartbeatThrottleIntervalMillis :: Word64

maxActivitiesPerSecond :: Maybe Double

maxTaskQueueActivitiesPerSecond :: Maybe Double

gracefulShutdownPeriodMillis :: Word64

nondeterminismAsWorkflowFail :: Bool

nondeterminismAsWorkflowFailForTypes :: [Text]

getWorkerConfig :: forall (ty :: WorkerType). Worker ty -> WorkerConfig

defaultWorkerConfig :: WorkerConfig

data WorkerError

Constructors

WorkerError

Fields

code :: WorkerErrorCode

message :: Text

data WorkerErrorCode

Constructors

SDKError

InitWorkerFailed

InitReplayWorkerFailed

InvalidProto

ReplayWorkerClosed

PollShutdown

PollFailure

CompletionFailure

InvalidWorkerConfig

UnknownError Word8

data WorkerType

Constructors

Real

Replay

data SWorkerType (ty :: WorkerType) where

A singleton type for WorkerType.

Constructors

SReal :: SWorkerType 'Real

SReplay :: SWorkerType 'Replay

data SomeWorkerType where

Constructors

SomeWorkerType :: forall (ty :: WorkerType). SWorkerType ty -> SomeWorkerType

newWorker :: Client -> WorkerConfig -> IO (Either WorkerError (Worker 'Real))

validateWorker :: Worker 'Real -> IO (Either WorkerValidationError ())

newReplayWorker :: Runtime -> WorkerConfig -> IO (Either WorkerError (Worker 'Replay, HistoryPusher))

closeWorker :: forall (ty :: WorkerType). Worker ty -> IO ()

type family InactiveForReplay (ty :: WorkerType) a where ...

Equations

InactiveForReplay 'Real a = a

InactiveForReplay 'Replay _1 = ()

data WorkflowActivation

Fields :

runId :: Lens' WorkflowActivation Data.Text.Text

timestamp :: Lens' WorkflowActivation Proto.Google.Protobuf.Timestamp.Timestamp

maybe'timestamp :: Lens' WorkflowActivation (Prelude.Maybe Proto.Google.Protobuf.Timestamp.Timestamp)

isReplaying :: Lens' WorkflowActivation Prelude.Bool

historyLength :: Lens' WorkflowActivation Data.Word.Word32

jobs :: Lens' WorkflowActivation [WorkflowActivationJob]

vec'jobs :: Lens' WorkflowActivation (Data.Vector.Vector WorkflowActivationJob)

availableInternalFlags :: Lens' WorkflowActivation [Data.Word.Word32]

vec'availableInternalFlags :: Lens' WorkflowActivation (Data.Vector.Unboxed.Vector Data.Word.Word32)

historySizeBytes :: Lens' WorkflowActivation Data.Word.Word64

continueAsNewSuggested :: Lens' WorkflowActivation Prelude.Bool

buildIdForCurrentTask :: Lens' WorkflowActivation Data.Text.Text

pollWorkflowActivation :: forall (ty :: WorkerType). KnownWorkerType ty => Worker ty -> IO (Either WorkerError WorkflowActivation)

data ActivityTask

Fields :

taskToken :: Lens' ActivityTask Data.ByteString.ByteString

maybe'variant :: Lens' ActivityTask (Prelude.Maybe ActivityTask'Variant)

maybe'start :: Lens' ActivityTask (Prelude.Maybe Start)

start :: Lens' ActivityTask Start

maybe'cancel :: Lens' ActivityTask (Prelude.Maybe Cancel)

cancel :: Lens' ActivityTask Cancel

pollActivityTask :: forall (ty :: WorkerType). KnownWorkerType ty => Worker ty -> IO (Either WorkerError ActivityTask)

data WorkflowActivationCompletion

Fields :

runId :: Lens' WorkflowActivationCompletion Data.Text.Text

maybe'status :: Lens' WorkflowActivationCompletion (Prelude.Maybe WorkflowActivationCompletion'Status)

maybe'successful :: Lens' WorkflowActivationCompletion (Prelude.Maybe Success)

successful :: Lens' WorkflowActivationCompletion Success

maybe'failed :: Lens' WorkflowActivationCompletion (Prelude.Maybe Failure)

failed :: Lens' WorkflowActivationCompletion Failure

completeWorkflowActivation :: forall (ty :: WorkerType). KnownWorkerType ty => Worker ty -> WorkflowActivationCompletion -> IO (Either WorkerError ())

data ActivityTaskCompletion

Fields :

taskToken :: Lens' ActivityTaskCompletion Data.ByteString.ByteString

result :: Lens' ActivityTaskCompletion Proto.Temporal.Sdk.Core.ActivityResult.ActivityResult.ActivityExecutionResult

maybe'result :: Lens' ActivityTaskCompletion (Prelude.Maybe Proto.Temporal.Sdk.Core.ActivityResult.ActivityResult.ActivityExecutionResult)

completeActivityTask :: forall (ty :: WorkerType). KnownWorkerType ty => Worker ty -> ActivityTaskCompletion -> IO (Either WorkerError ())

data ActivityHeartbeat

Fields :

taskToken :: Lens' ActivityHeartbeat Data.ByteString.ByteString

details :: Lens' ActivityHeartbeat [Proto.Temporal.Api.Common.V1.Message.Payload]

vec'details :: Lens' ActivityHeartbeat (Data.Vector.Vector Proto.Temporal.Api.Common.V1.Message.Payload)

recordActivityHeartbeat :: forall (ty :: WorkerType). KnownWorkerType ty => Worker ty -> ActivityHeartbeat -> IO (Either WorkerError ())

requestWorkflowEviction :: forall (ty :: WorkerType). KnownWorkerType ty => Worker ty -> RunId -> IO ()

initiateShutdown :: forall (ty :: WorkerType). KnownWorkerType ty => Worker ty -> IO ()

Initiate shutdown.

finalizeShutdown :: forall (ty :: WorkerType). KnownWorkerType ty => Worker ty -> IO (Either WorkerError ())

Completes shutdown and frees all resources. You should avoid simply dropping workers, as this does not allow async tasks to report any panics that may have occurred cleanly.

This should be called only after initiateShutdown has resolved and/or both polling functions have returned ShutDown errors.

data HistoryPusher

data History

Fields :

events :: Lens' History [HistoryEvent]

vec'events :: Lens' History (Data.Vector.Vector HistoryEvent)

pushHistory :: HistoryPusher -> WorkflowId -> Either ByteString History -> IO (Either WorkerError ())

closeHistory :: HistoryPusher -> IO ()

class KnownWorkerType (ty :: WorkerType) where

Methods

knownWorkerType :: SWorkerType ty