Temporal.Runtime

Documentation

data Runtime

A handle to the Rust Tokio runtime and thread-pool.

You almost always should use the globalRuntime value, which is initialized once for the entire process.

data TelemetryOptions

Constructors

OtelTelemetryOptions

Fields

url :: Text

headers :: Map Text Text

metricPeriodicity :: Maybe Periodicity

globalTags :: Map Text Text

PrometheusTelemetryOptions

Fields

socketAddr :: Text

globalTags :: Map Text Text

countersTotalSuffix :: Bool

unitSuffix :: Bool

NoTelemetry

data Periodicity

Constructors

Periodicity

Fields

seconds :: Word64

nanoseconds :: Word32

initializeRuntime :: TelemetryOptions -> IO Runtime

Initialize the Rust runtime and thread-pool.

withRuntime :: Runtime -> (Ptr Runtime -> IO a) -> IO a

Access the underlying Runtime pointer for calling out to Rust.

fetchLogs :: Runtime -> IO (Vector CoreLog)

The Rust runtime exports logs to the Haskell runtime. This function fetches those logs so they can be fed through a logging framework.

data CoreLog

Constructors

CoreLog

Fields

target :: Text

message :: Text

timestamp :: Value

level :: LogLevel

fields :: Object

spanContexts :: [Text]

data LogLevel

Constructors

Trace

Debug

Info

Warn

Error